apps/edge-agent/requirements.txt
requests==2.32.3
websocket-client==1.8.0
psutil==6.1.0
